sdk/python/pyproject.toml
[tool.poetry]
name = "sentience"
version = "0.0.1"
description = ""
authors = [
    "Kaspar Peterson <kaspar@galadriel.com>",
    "Kristjan Peterson <kristjan@galadriel.com>",
    "Krešimir Grofelnik <kresimir@galadriel.com>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.55.3"
pynacl = "^1.5.0"
requests = "2.31.0"
black = {version = "^24.8.0", optional = true}
pytest = {version = "^8.3.2", optional = true}
pytest-asyncio = {version = "0.24.0", optional = true}
pytest-mock = {version = "^3.14.0", optional = true}


[tool.poetry.extras]
dev = ["black", "pytest", "pytest-asyncio", "pytest-mock"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

#[tool.poetry.scripts]
# galadriel = "galadriel_node.cli.app:app"

[tool.pytest.ini_options]
asyncio_mode = "auto"